requirements.txt
APScheduler==3.10.1
asyncio==3.4.3
colorlog==6.7.0
edge-tts
psutil
fuzzywuzzy==0.18.0
httpx
openpyxl==3.1.2
picimagesearch
Pillow==10.0.0
PyYAML
regex
requests==2.28.2
soundfile==0.12.1
websockets==10.4
yiri-mirai
zhipuai==1.0.7
google-generativeai
openai
anthropic
bingart
emoji
num_thai
jamo
ko_pron
cn2an
opencc
indic_transliteration
jieba
pypinyin
unidecode
pyopenjtalk
inflect
eng_to_ipa
